refactor: rename kit util file
